voting section improvements
