+++ apps/site/src/components/pages/wordpress/voting/vote-email-input.tsx
@@ -1,9 +1,9 @@
 import { FormEvent, FunctionComponent, useRef, useState } from "react";
 
-import { apiClient } from "../../../lib/api-client";
-import { useEmailTextField } from "../../hooks/use-email-text-field";
-import { Input } from "./input";
-import { ApplicationIds } from "./request-another-application";
+import { apiClient } from "../../../../lib/api-client";
+import { useEmailTextField } from "../../../hooks/use-email-text-field";
+import { Input } from "../input";
+import { ApplicationIds } from "../voting/applications";
 
 const submitErrorText = "There was an error submitting your email";
